bug-hunt #111: check-full audit - fix all compile warnings under -Wpedantic -Wconversion -Wformat=2
Fixes across 50+ files to achieve zero warnings under check-full
(-Werror -Wpedantic -Wconversion -Wshadow -Wformat=2 -Wcast-align
 -Wstrict-prototypes -Wold-style-definition -Wmisleading-indentation).

Bug classes fixed:
  - Wpointer-arith: void* arithmetic in net code (net.c, pstore.c)
  - Wpedantic label/declaration: case label followed by decl (ipv6_core.c)
  - Wpedantic zero-size arrays: [0] → [1] in struct overlay (libc.h,
    virtio.h, cramfs.c)
  - Wpedantic empty initializer braces: {} → {0} (kunit test sentinels)
  - Wpedantic static forward decl: unsized static arrays → sized
    (kunit test arrays)
  - Wpedantic GCC extensions: ?: elvis operator → explicit ternary
    (dns_server.c); ({ }) statement expr. → inline function (div64.h)
  - Wpedantic void*/funcptr cast: used union cast (search.c)
  - Wpedantic func ptr type mismatch: pass NULL directly (service_proxy.c)
  - Wconversion narrowing (40+ files): uint64_t→uint32_t, int→uint8_t,
    size_t→int, unsigned→bitfield, etc.
  - Wformat=2 specifier mismatches (20+ files): %zu→%llu, %lu→%llu
    for uint64_t=unsigned_long_long on x86_64 freestanding GCC
  - Wnonnull: suppress for deliberate NULL-safety tests (test.c)
  - Missing generated header dependency: added -Ibuild for kpti.o
    in check-full target (Makefile)

+++ src/container/checkpoint.c
@@ -724,7 +724,7 @@ static int container_checkpoint_load(const char *save_path,
 
     kprintf("[Checkpoint] Checkpoint state loaded from %s "
             "(%u bytes header + %zu bytes state)\n",
-            save_path, sizeof(hdr), sizeof(*state_out));
+            save_path, (unsigned int)sizeof(hdr), sizeof(*state_out));
     return 0;
 }
