+++ src/drivers/pcie_aer.c
@@ -65,7 +65,7 @@ static int aer_handle_correctable(int bus, int dev, int func, uint32_t status)
             bus, dev, func, status);
 
     /* Clear status (in real hw, write 1 to clear) */
-    uint16_t aer_cap = pci_find_ext_cap(bus, dev, func, 0x0001);
+    uint16_t aer_cap = (uint16_t)pci_find_ext_cap(bus, dev, func, 0x0001);
     if (aer_cap) {
         /* pci_write32(bus, dev, func, aer_cap + PCI_ERR_COR_STATUS, status); */
     }
@@ -91,7 +91,7 @@ static int aer_handle_uncorrectable(int bus, int dev, int func, uint32_t status)
 /* Scan for AER capability on a device */
 static int aer_probe_device(int bus, int dev, int func)
 {
-    uint16_t aer_cap = pci_find_ext_cap(bus, dev, func, 0x0001); /* PCI_EXT_CAP_ID_AER */
+    uint16_t aer_cap = (uint16_t)pci_find_ext_cap(bus, dev, func, 0x0001); /* PCI_EXT_CAP_ID_AER */
     if (aer_cap) {
         kprintf("[AER] Found AER on %02x:%02x.%x (cap=0x%04x)\n",
                 bus, dev, func, aer_cap);
